# perf: not render popup on init