初始化的时候，在不需要完全加载Popup的情况下，不加载Popup，以减少初始render所需要的时间。
这个更新已经在3.x里已经有了，但是在2.x里没有，所以加过来。
PS：3.x只用于antd的next分支及6.0的版本，2.x用于antd 5.x，当前antd6处于alpha1版本，我在5.x上遇到了这个性能问题
参考：
https://github.com/ant-design/ant-design/pull/53844
https://github.com/react-component/trigger/commit/052b2ae8b067daa0ab25527c9242dfa50d827aca

本地测试的情况：
机器：Mac M4 Pro 24G
测试方法：某个组件内包含Tooltip 150个，计算渲染时间
结果：
chrome 默认速度：4.5ms -> 3.2ms
chrome 6x slowdown：29ms -> 22ms
整体提升约25%